api/types: move system info types to api/types/system
