+++ daemon/licensing.go
@@ -1,10 +1,10 @@
 package daemon // import "github.com/docker/docker/daemon"
 
 import (
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/system"
 	"github.com/docker/docker/dockerversion"
 )
 
-func (daemon *Daemon) fillLicense(v *types.Info) {
+func (daemon *Daemon) fillLicense(v *system.Info) {
 	v.ProductLicense = dockerversion.DefaultProductLicense
 }
