+++ daemon/licensing_test.go
@@ -3,13 +3,13 @@ package daemon // import "github.com/docker/docker/daemon"
 import (
 	"testing"
 
-	"github.com/docker/docker/api/types"
+	"github.com/docker/docker/api/types/system"
 	"github.com/docker/docker/dockerversion"
 	"gotest.tools/v3/assert"
 )
 
 func TestFillLicense(t *testing.T) {
-	v := &types.Info{}
+	v := &system.Info{}
 	d := &Daemon{
 		root: "/var/lib/docker/",
 	}
